+++ src/main/java/racingcar/io/writer/RacingWriter.java
@@ -1,4 +1,4 @@
-package racingcar.writer;
+package racingcar.io.writer;
 
 import java.util.List;
 import racingcar.model.ProceedResult;
