+++ src/test/java/racingcar/io/reader/RacingInfoReaderTest.java
@@ -1,4 +1,4 @@
-package racingcar.reader;
+package racingcar.io.reader;
 
 import camp.nextstep.edu.missionutils.Console;
 import org.junit.jupiter.api.AfterEach;
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InOrder;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import racingcar.ConsoleDisplay;
+import racingcar.io.ConsoleDisplay;
 import racingcar.model.RacingInfo;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
