+++ src/test/java/racingcar/io/reader/TurnValueReaderTest.java
@@ -1,4 +1,4 @@
-package racingcar.reader;
+package racingcar.io.reader;
 
 import camp.nextstep.edu.missionutils.Console;
 import org.junit.jupiter.api.AfterEach;
@@ -6,8 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
-import racingcar.ConsoleDisplay;
-import racingcar.reader.TurnValueReader;
+import racingcar.io.ConsoleDisplay;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
