+++ src/test/java/racingcar/io/writer/ProceedResultWriterTest.java
@@ -1,9 +1,9 @@
-package racingcar.writer;
+package racingcar.io.writer;
 
 import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.ConsoleDisplay;
+import racingcar.io.ConsoleDisplay;
 import racingcar.model.ProceedResult;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
